+++ lib/eve_online.rb
@@ -66,6 +66,7 @@ require 'eve_online/esi/universe_type'
 require 'eve_online/esi/universe_regions'
 require 'eve_online/esi/universe_region'
 require 'eve_online/esi/universe_constellations'
+require 'eve_online/esi/universe_constellation'
 
 require 'eve_online/esi/models/base'
 require 'eve_online/esi/models/loyalty_point'
@@ -110,6 +111,7 @@ require 'eve_online/esi/models/bloodline'
 require 'eve_online/esi/models/faction'
 require 'eve_online/esi/models/ancestry'
 require 'eve_online/esi/models/region'
+require 'eve_online/esi/models/constellation'
 
 module EveOnline
 end